+++ main.tf
@@ -5,11 +5,11 @@ locals {
   public_subnet_ids  = coalescelist(module.vpc.public_subnets, var.public_subnet_ids, [""])
 
   # Atlantis
-  atlantis_image = var.atlantis_image == "" ? "runatlantis/atlantis:${var.atlantis_version}" : var.atlantis_image
+  atlantis_image = var.atlantis_image == "" ? "ghcr.io/runatlantis/atlantis:${var.atlantis_version}" : var.atlantis_image
   atlantis_url = "https://${coalesce(
     var.atlantis_fqdn,
     element(concat(aws_route53_record.atlantis.*.fqdn, [""]), 0),
-    module.alb.this_lb_dns_name,
+    module.alb.lb_dns_name,
     "_"
   )}"
   atlantis_url_events = "${local.atlantis_url}/events"
@@ -68,14 +68,17 @@ locals {
       value = var.atlantis_bitbucket_base_url
     },
     {
-      name  = "ATLANTIS_REPO_WHITELIST"
-      value = join(",", var.atlantis_repo_whitelist)
+      name  = "ATLANTIS_REPO_ALLOWLIST"
+      value = join(",", var.atlantis_repo_allowlist)
     },
     {
       name  = "ATLANTIS_HIDE_PREV_PLAN_COMMENTS"
       value = var.atlantis_hide_prev_plan_comments
     },
   ]
+
+  # ECS task definition
+  latest_task_definition_rev = var.external_task_definition_updates ? max(aws_ecs_task_definition.atlantis.revision, data.aws_ecs_task_definition.atlantis[0].revision) : aws_ecs_task_definition.atlantis.revision
 
   # Secret access tokens
   container_definition_secrets_1 = local.secret_name_key != "" && local.secret_name_value_from != "" ? [
@@ -99,7 +102,11 @@ locals {
     },
     var.tags,
   )
+
+  policies_arn = var.policies_arn != null ? var.policies_arn : ["arn:${data.aws_partition.current.partition}:iam::aws:policy/service-role/AmazonECSTaskExecutionRolePolicy"]
 }
+
+data "aws_partition" "current" {}
 
 data "aws_region" "current" {}
 
@@ -110,9 +117,9 @@ data "aws_route53_zone" "this" {
   private_zone = false
 }
 
-###################
+################################################################################
 # Secret for webhook
-###################
+################################################################################
 resource "random_id" "webhook" {
   count = var.atlantis_github_webhook_secret != "" ? 0 : 1
 
@@ -159,12 +166,13 @@ resource "aws_ssm_parameter" "atlantis_bitbucket_user_token" {
   tags = local.tags
 }
 
-###################
+################################################################################
 # VPC
-###################
+################################################################################
+
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "v2.64.0"
+  version = "v3.6.0"
 
   create_vpc = var.vpc_id == ""
 
@@ -181,19 +189,19 @@ module "vpc" {
   tags = local.tags
 }
 
-###################
+################################################################################
 # ALB
-###################
+################################################################################
 module "alb" {
   source  = "terraform-aws-modules/alb/aws"
-  version = "v5.10.0"
+  version = "v6.5.0"
 
   name     = var.name
   internal = var.internal
 
   vpc_id          = local.vpc_id
   subnets         = local.public_subnet_ids
-  security_groups = flatten([module.alb_https_sg.this_security_group_id, module.alb_http_sg.this_security_group_id, var.security_group_ids])
+  security_groups = flatten([module.alb_https_sg.security_group_id, module.alb_http_sg.security_group_id, var.security_group_ids])
 
   access_logs = {
     enabled = var.alb_logging_enabled
@@ -201,13 +209,17 @@ module "alb" {
     prefix  = var.alb_log_location_prefix
   }
 
+  enable_deletion_protection = var.alb_enable_deletion_protection
+
+  drop_invalid_header_fields = var.alb_drop_invalid_header_fields
+
   listener_ssl_policy_default = var.alb_listener_ssl_policy_default
   https_listeners = [
     {
       target_group_index   = 0
       port                 = 443
       protocol             = "HTTPS"
-      certificate_arn      = var.certificate_arn == "" ? module.acm.this_acm_certificate_arn : var.certificate_arn
+      certificate_arn      = var.certificate_arn == "" ? module.acm.acm_certificate_arn : var.certificate_arn
       action_type          = local.alb_authenication_method
       authenticate_oidc    = var.alb_authenticate_oidc
       authenticate_cognito = var.alb_authenticate_cognito
@@ -234,6 +246,9 @@ module "alb" {
       backend_port         = var.atlantis_port
       target_type          = "ip"
       deregistration_delay = 10
+      health_check = {
+        path = "/healthz"
+      }
     },
   ]
 
@@ -269,12 +284,31 @@ resource "aws_lb_listener_rule" "unauthenticated_access_for_cidr_blocks" {
   }
 }
 
-###################
+# Forward action for certain URL paths to bypass authentication (eg. GitHub webhooks)
+resource "aws_lb_listener_rule" "unauthenticated_access_for_webhook" {
+  count = var.allow_unauthenticated_access && var.allow_github_webhooks ? 1 : 0
+
+  listener_arn = module.alb.https_listener_arns[0]
+  priority     = var.allow_unauthenticated_webhook_access_priority
+
+  action {
+    type             = "forward"
+    target_group_arn = module.alb.target_group_arns[0]
+  }
+
+  condition {
+    path_pattern {
+      values = ["/events"]
+    }
+  }
+}
+
+################################################################################
 # Security groups
-###################
+################################################################################
 module "alb_https_sg" {
   source  = "terraform-aws-modules/security-group/aws//modules/https-443"
-  version = "v3.17.0"
+  version = "v4.3.0"
 
   name        = "${var.name}-alb-https"
   vpc_id      = local.vpc_id
@@ -287,7 +321,7 @@ module "alb_https_sg" {
 
 module "alb_http_sg" {
   source  = "terraform-aws-modules/security-group/aws//modules/http-80"
-  version = "v3.17.0"
+  version = "v4.3.0"
 
   name        = "${var.name}-alb-http"
   vpc_id      = local.vpc_id
@@ -300,7 +334,7 @@ module "alb_http_sg" {
 
 module "atlantis_sg" {
   source  = "terraform-aws-modules/security-group/aws"
-  version = "v3.17.0"
+  version = "v4.3.0"
 
   name        = var.name
   vpc_id      = local.vpc_id
@@ -312,7 +346,7 @@ module "atlantis_sg" {
       to_port                  = var.atlantis_port
       protocol                 = "tcp"
       description              = "Atlantis"
-      source_security_group_id = module.alb_https_sg.this_security_group_id
+      source_security_group_id = module.alb_https_sg.security_group_id
     },
   ]
 
@@ -321,12 +355,12 @@ module "atlantis_sg" {
   tags = merge(local.tags, var.atlantis_security_group_tags)
 }
 
-###################
+################################################################################
 # ACM (SSL certificate)
-###################
+################################################################################
 module "acm" {
   source  = "terraform-aws-modules/acm/aws"
-  version = "v2.12.0"
+  version = "v3.2.0"
 
   create_certificate = var.certificate_arn == ""
 
@@ -337,9 +371,9 @@ module "acm" {
   tags = local.tags
 }
 
-###################
+################################################################################
 # Route53 record
-###################
+################################################################################
 resource "aws_route53_record" "atlantis" {
   count = var.create_route53_record ? 1 : 0
 
@@ -348,27 +382,29 @@ resource "aws_route53_record" "atlantis" {
   type    = "A"
 
   alias {
-    name                   = module.alb.this_lb_dns_name
-    zone_id                = module.alb.this_lb_zone_id
+    name                   = module.alb.lb_dns_name
+    zone_id                = module.alb.lb_zone_id
     evaluate_target_health = true
   }
 }
 
-###################
+################################################################################
 # ECS
-###################
+################################################################################
 module "ecs" {
   source  = "terraform-aws-modules/ecs/aws"
-  version = "v2.5.0"
+  version = "v3.3.0"
 
   name               = var.name
   container_insights = var.ecs_container_insights
 
   capacity_providers = ["FARGATE", "FARGATE_SPOT"]
 
-  default_capacity_provider_strategy = {
-    capacity_provider = var.ecs_fargate_spot ? "FARGATE_SPOT" : "FARGATE"
-  }
+  default_capacity_provider_strategy = [
+    {
+      capacity_provider = var.ecs_fargate_spot ? "FARGATE_SPOT" : "FARGATE"
+    }
+  ]
 
   tags = local.tags
 }
@@ -385,6 +421,15 @@ data "aws_iam_policy_document" "ecs_tasks" {
       type        = "Service"
       identifiers = compact(distinct(concat(["ecs-tasks.amazonaws.com"], var.trusted_principals)))
     }
+
+    dynamic "principals" {
+      for_each = length(var.trusted_entities) > 0 ? [true] : []
+
+      content {
+        type        = "AWS"
+        identifiers = var.trusted_entities
+      }
+    }
   }
 }
 
@@ -397,10 +442,10 @@ resource "aws_iam_role" "ecs_task_execution" {
 }
 
 resource "aws_iam_role_policy_attachment" "ecs_task_execution" {
-  count = length(var.policies_arn)
+  for_each = toset(local.policies_arn)
 
   role       = aws_iam_role.ecs_task_execution.id
-  policy_arn = element(var.policies_arn, count.index)
+  policy_arn = each.value
 }
 
 # ref: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/specifying-sensitive-data.html
@@ -455,7 +500,7 @@ resource "aws_iam_role_policy" "ecs_task_access_secrets" {
 
 module "container_definition_github_gitlab" {
   source  = "cloudposse/ecs-container-definition/aws"
-  version = "v0.45.2"
+  version = "v0.58.1"
 
   container_name  = var.name
   container_image = local.atlantis_image
@@ -512,7 +557,7 @@ module "container_definition_github_gitlab" {
 
 module "container_definition_bitbucket" {
   source  = "cloudposse/ecs-container-definition/aws"
-  version = "v0.45.2"
+  version = "v0.58.1"
 
   container_name  = var.name
   container_image = local.atlantis_image
@@ -577,10 +622,19 @@ resource "aws_ecs_task_definition" "atlantis" {
 
   container_definitions = local.container_definitions
 
+  dynamic "ephemeral_storage" {
+    for_each = var.enable_ephemeral_storage ? [1] : []
+    content {
+      size_in_gib = var.ephemeral_storage_size
+    }
+  }
+
   tags = local.tags
 }
 
 data "aws_ecs_task_definition" "atlantis" {
+  count = var.external_task_definition_updates ? 1 : 0
+
   task_definition = var.name
 
   depends_on = [aws_ecs_task_definition.atlantis]
@@ -588,20 +642,20 @@ data "aws_ecs_task_definition" "atlantis" {
 
 resource "aws_ecs_service" "atlantis" {
   name    = var.name
-  cluster = module.ecs.this_ecs_cluster_id
-  task_definition = "${data.aws_ecs_task_definition.atlantis.family}:${max(
-    aws_ecs_task_definition.atlantis.revision,
-    data.aws_ecs_task_definition.atlantis.revision,
-  )}"
+  cluster = module.ecs.ecs_cluster_id
+
+  task_definition                    = "${var.name}:${local.latest_task_definition_rev}"
   desired_count                      = var.ecs_service_desired_count
   launch_type                        = var.ecs_fargate_spot ? null : "FARGATE"
   platform_version                   = var.ecs_service_platform_version
   deployment_maximum_percent         = var.ecs_service_deployment_maximum_percent
   deployment_minimum_healthy_percent = var.ecs_service_deployment_minimum_healthy_percent
+  force_new_deployment               = var.ecs_service_force_new_deployment
+  enable_execute_command             = var.ecs_service_enable_execute_command
 
   network_configuration {
     subnets          = local.private_subnet_ids
-    security_groups  = [module.atlantis_sg.this_security_group_id]
+    security_groups  = [module.atlantis_sg.security_group_id]
     assign_public_ip = var.ecs_service_assign_public_ip
   }
 
@@ -609,6 +663,15 @@ resource "aws_ecs_service" "atlantis" {
     container_name   = var.name
     container_port   = var.atlantis_port
     target_group_arn = element(module.alb.target_group_arns, 0)
+  }
+
+  dynamic "load_balancer" {
+    for_each = var.extra_load_balancers
+    content {
+      container_name   = load_balancer.value["container_name"]
+      container_port   = load_balancer.value["container_port"]
+      target_group_arn = load_balancer.value["target_group_arn"]
+    }
   }
 
   dynamic "capacity_provider_strategy" {
@@ -622,12 +685,12 @@ resource "aws_ecs_service" "atlantis" {
   enable_ecs_managed_tags = var.enable_ecs_managed_tags
   propagate_tags          = var.propagate_tags
 
-  tags = local.tags
+  tags = var.use_ecs_old_arn_format ? null : local.tags
 }
 
-###################
+################################################################################
 # Cloudwatch logs
-###################
+################################################################################
 resource "aws_cloudwatch_log_group" "atlantis" {
   name              = var.name
   retention_in_days = var.cloudwatch_log_retention_in_days