+++ modules/github-repository-webhook/outputs.tf
@@ -1,9 +1,10 @@
-output "this_repository_webhook_urls" {
+output "repository_webhook_urls" {
   description = "Webhook URL"
   value       = github_repository_webhook.this.*.url
 }
 
-output "this_repository_webhook_secret" {
+output "repository_webhook_secret" {
   description = "Webhook secret"
   value       = var.webhook_secret
+  sensitive   = true
 }